Add key props to sections and separate content item from content items
