+++ src/components/StageEditor/sections/ContentItems.js
@@ -1,25 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Button, MarkdownInput, FileInput } from '../../../components/Form';
-
-const contentTypes = {
-  text: MarkdownInput,
-  image: (value, onChange) => (
-    <FileInput value={value} onChange={onChange}>{ src => (<img src={src} alt="preview" />) }</FileInput>
-  ),
-  audio: FileInput,
-  video: FileInput,
-};
-
-const ContentItem = ({ type, content, onChange }) => {
-  const ContentType = contentTypes[type];
-
-  return (
-    <div styles={{ borderTop: '2px', borderColor: 'black' }}>
-      <ContentType value={content} onChange={value => onChange({ type, content: value })} />
-    </div>
-  );
-};
+import { Button } from '../../../components/Form';
+import ContentItem from './ContentItem';
 
 class ContentItems extends Component {
   static propTypes = {
@@ -37,19 +19,17 @@ class ContentItems extends Component {
   };
 
   createNewItem = (type) => {
-    console.log('NEW ITEM', this.props.stage.contentItems, type);
     const contentItems = this.props.stage.contentItems || [];
     this.props.onChange({ contentItems: [...contentItems, { type, content: undefined }] });
   }
 
   updateItem = (newItem, index) => {
-    console.log('UPDATE ITEM', this.props.stage.contentItems, newItem, index);
     const contentItems = this.props.stage.contentItems
       .map((item, i) => {
         if (i !== index) { return item; }
         return newItem;
       });
-    console.log({ contentItems });
+
     this.props.onChange({ contentItems });
   }
 
@@ -59,7 +39,7 @@ class ContentItems extends Component {
     } = this.props;
 
     return ([
-      <div className="edit-stage__section">
+      <div className="edit-stage__section" key="edit">
         { contentItems &&
           contentItems.map(
             (props, index) => (
@@ -72,13 +52,13 @@ class ContentItems extends Component {
           )
         }
         <div>
-          <Button onClick={() => this.createNewItem('copy')}>Add copy</Button>
+          <Button onClick={() => this.createNewItem('text')}>Add copy</Button>
           <Button onClick={() => this.createNewItem('image')}>Add image</Button>
           <Button onClick={() => this.createNewItem('audio')}>Add audio</Button>
           <Button onClick={() => this.createNewItem('video')}>Add video</Button>
         </div>
       </div>,
-      <div className="edit-skip-logic__guidance">
+      <div className="edit-skip-logic__guidance" key="guidance">
         Add your content here
       </div>,
     ]);
